*refactoring of ConvertSplitNode.
Move ConvertSplitNode from OnnxRuntime.h into ConvertSplitNode.cpp
and replace asserts with SYNET_ERROR messages.
